requirements.txt
vncorenlp
datasets==1.0.2
transformers==4.8.0
sacrebleu==1.4.12
rouge_score
PyYAML
Flask==2.0.2
streamlit